+++ src/components/Asset/Edit/EditFeedback.module.css
@@ -5,6 +5,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
+  flex-direction: column;
 }
 
 .feedback h3 {
